requirements.txt
flask
tensorflow==2.13.0
pillow
requests        
gunicorn        